backend/pyproject.toml
[project]
name = "miro-chiral-backend"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.4.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "openai>=1.50.0",
    "anthropic>=0.35.0",
    "httpx>=0.27.0",
    "graphiti-core>=0.3.0",
    "neo4j>=5.24.0",
    "camel-ai[oasis]>=0.2.0",
    "pypdf>=5.0.0",
    "sentence-transformers>=3.2.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "transformers>=4.40.0",
    "networkx>=3.0",
    "trafilatura>=2.0.0",
    "duckduckgo-search>=6.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
